+++ src/RMC.cpp
@@ -46,7 +46,7 @@ extern "C" {
 		
 		//Independent variables
 		vector< AD<double> > params( par_info.nPars);
-		for( size_t i=0; i<par_info.nPars; i++)
+		for( int i=0; i<par_info.nPars; i++)
 			params.at( i) = pars[i];
 		CppAD::Independent( params);
 		
@@ -68,7 +68,7 @@ extern "C" {
 			lbfgsminimize( par_info.nPars, *m, x, *epsg, *epsf, *epsx, *max_iter, *R_conv);
 		
 		vector<double> esti;
-		for( size_t i=0; i<par_info.nPars; i++){
+		for( int i=0; i<par_info.nPars; i++){
 			esti.push_back( x(i+1));
 			pars[i] = x( i+1);
 		}		
@@ -80,9 +80,9 @@ extern "C" {
 		if( *R_calcHes==1)
 			temphes = F.Hessian( esti, 0);
 
-		for( size_t i=0; i<par_info.nPars; i++){
+		for( int i=0; i<par_info.nPars; i++){
 			R_scores[i] = tempScore.at( i);
-			for( size_t j=0; j<par_info.nPars; j++)
+			for( int j=0; j<par_info.nPars; j++)
 				R_hessian[j*par_info.nPars+i] = temphes.at( j*par_info.nPars+i);
 		}
 	
@@ -97,14 +97,14 @@ void funcgrad(ap::real_1d_array x, double &f, ap::real_1d_array &g)
 	vector<double> gtemp( dim);
 	
 	
-	for( size_t i=1; i<=dim; i++)
+	for( int i=1; i<=dim; i++)
 		newX.push_back( x(i));
 		
 	ftemp = F.Forward( 0, newX);
 	gtemp = F.RevOne( newX, 0);
 	
 	f = 0 - ftemp.at( 0);
-	for( size_t i=0; i<dim; i++)
+	for( int i=0; i<dim; i++)
 		g(i+1) = 0 - gtemp.at( i);
 }
 
@@ -119,7 +119,7 @@ vector< AD<double> > ploglike( const vector< AD<double> > &params1, const RMCdat
 	int start = 0;
 	AD<double> penTerm = 0.0;
 	int stop = dat1.nObs.at( 0);
-	for( size_t i=0; i<dat1.nChains; i++){
+	for( int i=0; i<dat1.nChains; i++){
 		templl.at( i) = 0;
 		templl.at( i) = chainploglike( params1, dat1, par_info1, start, stop);
 
@@ -128,10 +128,10 @@ vector< AD<double> > ploglike( const vector< AD<double> > &params1, const RMCdat
 			stop += dat1.nObs.at( i+1);
 		}
 	}
-	for( size_t i=0; i<dat1.nChains; i++)
+	for( int i=0; i<dat1.nChains; i++)
 		pll1.at( 0) += templl.at( i);
 	if( penalty1>0) {
-		for( size_t i=0; i<par_info1.nPars; i++)
+		for( int i=0; i<par_info1.nPars; i++)
 			penTerm += params1.at( i)*params1.at( i) / ( 2 * penalty1);
 		pll1.at( 0) -= penTerm;
 //		Rprintf( "%f", (double)penTerm);
@@ -147,10 +147,10 @@ AD<double> first_logl( const vector< AD<double> > &params5, const RMCdata &dat5,
 	AD<double> sum = 0.;
 	
 	piVec5 = calcPiVec( params5, dat5, par_info5, index5);
-	for( size_t j=0; j<dat5.nCats; j++)
+	for( int j=0; j<dat5.nCats; j++)
 		phiVec5.at( j) = calcSinglePhi( params5, dat5, par_info5, index5, j+1);
 	
-	for( size_t i=0; i<dat5.nCats; i++){
+	for( int i=0; i<dat5.nCats; i++){
 		dist.at( i) = piVec5.at( i) / phiVec5.at( i);
 		sum += dist.at( i);
 	}
@@ -164,7 +164,7 @@ AD<double> chainploglike( const vector< AD<double> > &params2, const RMCdata &da
 	vector< AD<double> > piVec( dat2.nCats, 0), phiVec( dat2.nCats, 0), prevDist( dat2.nCats, 0), newDist( dat2.nCats, 0);
 	
 	pllsum += first_logl( params2, dat2, par_info2, start2);
-	for( size_t ii=start2+1; ii<stop2; ii++){
+	for( int ii=start2+1; ii<stop2; ii++){
 		temp = 0;
 		piVec = calcPiVec( params2, dat2, par_info2, ii);		//used to be ii-1
 		if( dat2.states.at( ii) != -9){							//current observation IS NOT -9
@@ -177,7 +177,7 @@ AD<double> chainploglike( const vector< AD<double> > &params2, const RMCdata &da
 			}
 			else{																			//First observation after -9
 				temp = 0;
-				for( size_t jj=0; jj<dat2.nCats; jj++){
+				for( int jj=0; jj<dat2.nCats; jj++){
 					phi = calcSinglePhi( params2, dat2, par_info2, ii, jj+1);		//used to be ii-1
 					if( jj == dat2.states.at( ii)-1)
 						temp += ( ( 1-phi) + phi*piVec.at( dat2.states.at( ii)-1)) * prevDist.at( jj);
@@ -189,24 +189,24 @@ AD<double> chainploglike( const vector< AD<double> > &params2, const RMCdata &da
 		}
 		else{																				//current observation IS -9
 			if( dat2.states.at( ii-1) == -9){									//Internal -9
-				for( size_t jj=0; jj<dat2.nCats; jj++)
+				for( int jj=0; jj<dat2.nCats; jj++)
 					phiVec.at( jj) = calcSinglePhi( params2, dat2, par_info2, ii, jj+1);		//used to be ii-1
-				for( size_t jj=0; jj<dat2.nCats; jj++){
+				for( int jj=0; jj<dat2.nCats; jj++){
 					newDist.at( jj) = 0;
 					Tpel = 0;
-					for( size_t kk=0; kk<dat2.nCats; kk++){
+					for( int kk=0; kk<dat2.nCats; kk++){
 						Tpel = phiVec.at( kk) * piVec.at( jj);
 						if( jj == kk)
 							Tpel += 1-phiVec.at( kk);
 						newDist.at( jj) += Tpel * prevDist.at( kk);
 					}
 				}
-				for( size_t jj=0; jj<dat2.nCats; jj++)
+				for( int jj=0; jj<dat2.nCats; jj++)
 					prevDist.at( jj) = newDist.at( jj);
 			}
 			else{																	//First -9 in a block
 				phi = calcSinglePhi( params2, dat2, par_info2, ii, dat2.states.at( ii-1));		//used to be ii-1
-				for( size_t jj=0; jj<dat2.nCats; jj++){
+				for( int jj=0; jj<dat2.nCats; jj++){
 					prevDist.at( jj) = piVec.at( jj) * phi;
 					if( dat2.states.at( ii-1)-1 == jj)
 						prevDist.at( jj) += 1-phi;
@@ -224,7 +224,7 @@ vector< AD<double> > calcPiVec( const vector< AD<double> > &params3, const RMCda
 	vector< AD<double> > lp( dat3.nCats, 0), explp( dat3.nCats, 0);
 	AD<double> sumexplp=0;
 		
-	for( size_t ii=0; ii<dat3.nCats-1; ii++){
+	for( int ii=0; ii<dat3.nCats-1; ii++){
 		for( size_t jj=0; jj<par_info3.piIndex.size(); jj++)
 			lp.at( ii) += dat3.X.at( row).at( par_info3.piIndex.at( jj)) * params3.at( par_info3.nPhi + ii*par_info3.nPi + jj);
 		explp.at( ii) = exp( lp.at( ii));
@@ -232,7 +232,7 @@ vector< AD<double> > calcPiVec( const vector< AD<double> > &params3, const RMCda
 	}
 	
 	resVec.at( 0) = 1;
-	for( size_t ii=0; ii<dat3.nCats-1; ii++){
+	for( int ii=0; ii<dat3.nCats-1; ii++){
 		resVec.at( ii+1) = explp.at( ii) / ( 1+sumexplp);
 		resVec.at( 0) -= resVec.at( ii+1);
 	}
@@ -245,7 +245,7 @@ AD<double> calcSinglePhi( const vector< AD<double> > &params3, const RMCdata &da
 	AD<double> lp=0, phi=0;
 	int temp=0;
 	
-	for( size_t i=0; i<(fromState-1); i++)
+	for( int i=0; i<(fromState-1); i++)
 		temp += par_info3.phiIndex.at( i).size();
 	for( size_t jj=0; jj<par_info3.phiIndex.at( fromState-1).size(); jj++)
 		lp += dat3.X.at( row).at( par_info3.phiIndex.at( fromState-1).at( jj)) * params3.at( temp + jj);
@@ -268,14 +268,14 @@ void RMCdata::setVals( int R_nCats1, int R_nCovars1, int R_nTot1, int R_nChains1
 	nTot = R_nTot1;
 	nChains = R_nChains1;
 			
-	for( size_t i=0; i<nTot; i++){
+	for( int i=0; i<nTot; i++){
 		X.push_back( vector<double>());
-		for( size_t j=0; j<nCovars; j++)
+		for( int j=0; j<nCovars; j++)
 			X.at(i).push_back( R_X1[j*nTot+i]);
 		states.push_back( R_states1[i]);
 	}
 	
-	for( size_t i=0; i<nChains; i++)
+	for( int i=0; i<nChains; i++)
 		nObs.push_back( R_nObs1[i]);
 }
 
@@ -290,7 +290,7 @@ void PARdata::setVals( int *R_phiID1, int *R_piID1, int R_nCovars1, int R_nCats1
 	nPi = 0;
 	nPhi = 0;
 	
-	for( size_t i=0; i<R_nCovars1; i++){
+	for( int i=0; i<R_nCovars1; i++){
 		if( R_piID1[i] == 1){
 			nPi++;
 			piIndex.push_back( i);
