+++ src/predict.cpp
@@ -14,14 +14,14 @@ vector< AD<double> > statDist( const vector< AD<double> > &params5, const RMCdat
 	AD<double> sum = 0.;
 	
 	piVec5 = calcPiVec( params5, dat5, par_info5, index5);
-	for( size_t j=0; j<dat5.nCats; j++)
+	for( int j=0; j<dat5.nCats; j++)
 		phiVec5.at( j) = calcSinglePhi( params5, dat5, par_info5, index5, j+1);
 	
-	for( size_t i=0; i<dat5.nCats; i++){
+	for( int i=0; i<dat5.nCats; i++){
 		dist.at( i) = piVec5.at( i) / phiVec5.at( i);
 		sum += dist.at( i);
 	}
-	for( size_t i=0; i<dat5.nCats; i++)
+	for( int i=0; i<dat5.nCats; i++)
 		dist.at( i) = dist.at( i) / sum;
 
 	return( dist);	
@@ -43,7 +43,7 @@ extern "C" {
 	//Parameter vector - independent
 	vector< AD<double> > params( predPar_info.nPars);
 	vector<double> predVal;
-	for( size_t i=0; i<predPar_info.nPars; i++){
+	for( int i=0; i<predPar_info.nPars; i++){
 		params.at( i) = Rpars[i];
 		predVal.push_back( Rpars[i]);
 	}
@@ -54,15 +54,15 @@ extern "C" {
 	//prediction matrix - dependent (row wise)
 	vector< vector< AD<double> > > dist;
 
-	for( size_t i=0; i<predDat.nTot; i++)
+	for( int i=0; i<predDat.nTot; i++)
 		dist.push_back( statDist( params, predDat, predPar_info, i));
 		
 	AD<double> sum=0;
 	vector< AD<double> > areaPred;
 	
-	for( size_t i=0; i<predDat.nCats; i++){
+	for( int i=0; i<predDat.nCats; i++){
 		sum = 0;
-		for( size_t j=0; j<predDat.nTot; j++)
+		for( int j=0; j<predDat.nTot; j++)
 			sum += dist.at(j).at(i);
 		areaPred.push_back( sum / predDat.nTot);
 	}
@@ -74,13 +74,13 @@ extern "C" {
 	}
 
 	//"output" array
-	for( size_t i=0; i<predDat.nTot; i++)
-		for( size_t j=0; j<predDat.nCats; j++)
+	for( int i=0; i<predDat.nTot; i++)
+		for( int j=0; j<predDat.nCats; j++)
 			ppts[i+j*predDat.nTot] = Value( dist.at(i).at(j));
-	for( size_t i=0; i<predDat.nCats; i++)
+	for( int i=0; i<predDat.nCats; i++)
 		area[i] = Value( areaPred.at(i));
 	if( *RdoCov==1){
-		for( size_t i=0; i<predDat.nCats; i++)
+		for( int i=0; i<predDat.nCats; i++)
 			for( size_t j=0; j<params.size(); j++)
 				Rjac[j*predDat.nCats+i] = jac.at( i*params.size()+j);
 	}
